List Manipulation Preparation
This update includes preparation for the manipulation of lists in order to construct the heatmap. This included tidying up headers and file writers into nested methods so they're easier to just comment out 1 instead of 2 when testing. This also makes it easier to read what is going on, quickly. Wrote the list of champions to a txt file now, which doesn't have duplicates. This also increased the time that it takes the full program to run (unfortunately).