Massive File Construction Changes
Changed the way the files are being written and making sure that the list of champions is written in alphabetical order. This is done with a single while loop to try to make it as-fast-as-possible. The alphabetical list is needed in order to create the heatmap for the future steps.