File updates
Few improvement and getting closer to getting the entire csv file done and formatted. It now is able to keep the win / loss of a single game of a particular champion that was played with the character in X column. Problems now are something with list index not being in range-- going to find out how-- since it seems that the range should be the exact same  in the index being presented.