Massive Minimalizing and Preparing
Preparing the files to be able to be used more frequently and for a team setting. This also involved deleting previous files and inputting names of people in IC (Illinois College) team. As for the minimalizing, I decided to get rid of a bunch of unnecessary file directory links where it would point to something extremely absolute (this seemed to cause an error). There was more but I can't remember it right now.